requirements.txt
# Core Voice & Audio
vosk==0.3.45
pyaudio==0.2.14
pyttsx3==2.90

# Web Framework & GUI
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# Web Scraping & Requests
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Configuration & Environment
python-dotenv==1.0.0

# System Monitoring
psutil==5.9.6
watchdog==3.0.0

# Image Processing & OCR
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3

# Windows Specific
pywin32==306
pygetwindow==0.0.9
pyautogui==0.9.54
comtypes==1.2.0

# File Generation & Code Templates
jinja2==3.1.2
markupsafe==2.1.3
